+++ client/form.html
@@ -12,8 +12,6 @@
     <link rel="stylesheet" type="text/css" href="mapbox.css">
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
     integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
-    integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" 
     integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
     <script src="main.js"></script>
@@ -55,7 +53,7 @@
       <div style="padding: 20px;color:whitesmoke">
           <h1 style="text-align: center">Any problems in your area?
           Post them here!</h1><br>
-          <p style="text-align: center">(Allow access to location if prompted)</p>
+          <p style="text-align: center">(Drag Marker to set location. Allow access to location if prompted)</p>
       </div>
 
     <section>
@@ -156,7 +154,7 @@
                     
                         <div class="form-group" id="uploadimage">
 
-                                <label for="picfile" class="btn btn-secondary">Upload Image</label>
+                                <label for="picfile">Upload Image</label>
 
                                 <input type="file" class="form-control-file" id="picfile" name="image" required>
                               </div>
@@ -176,7 +174,7 @@
             <div class="row">
                             <div class="col-10 col-sm-10 ">
                                 <!-- <label>Location&emsp;&emsp;</label> -->
-                                <textarea style="height:70px " placeholder=" Enter Location" id="location" rows="2" cols="50" name="location" required></textarea>
+                                <textarea style="height:70px " placeholder=" Enter Landmark/Area/Colony" id="location" rows="2" cols="50" name="location" required></textarea>
                               
                             </div>
