+++ client/main.css
@@ -246,10 +246,6 @@ textarea,
   }
 }
 
-label.btn {
-  background-color: rgb(43, 161, 78);
-}
-
 @media only screen and (max-width: 768px) {
   .card h4 {
     font-size: 1rem;
