+++ server.js
@@ -2,14 +2,13 @@
 const express = require('express');
 const path = require('path');
 require('./db/mongoose');
-const passport = require("passport")
 const bodyParser = require('body-parser');
 const userRoutes = require('./routes/userRoutes');
 const adminRoutes = require('./routes/adminRoutes');
-const Admin = require("./models/adminModel");
-var LocalStrategy = require("passport-local");
-var middleware = require("./middleware/index");
-
+const Admin = require('./models/adminModel');
+const passport = require('passport');
+var LocalStrategy = require('passport-local');
+var middleware = require('./middleware/auth');
 
 const app = express();
 
@@ -21,11 +20,13 @@ app.use(bodyParser.json());
 app.use(express.static(path.join(__dirname, 'client')));
 
 //Passport configuration
-app.use(require("express-session")({
-  secret: "best website ever!",
-  resave: false,
-  saveUninitialized: false
-}));
+app.use(
+  require('express-session')({
+    secret: 'best website ever!',
+    resave: false,
+    saveUninitialized: false
+  })
+);
 app.use(passport.initialize());
 app.use(passport.session());
 passport.use(new LocalStrategy(Admin.authenticate()));
@@ -48,28 +49,20 @@ app.get('/form', (req, res, next) => {
   res.sendFile(__dirname + '/client/form.html');
 });
 
+app.get('/register', (req, res, next) => {
+  res.sendFile(__dirname + '/client/register.html');
+});
+
+app.get('/login', (req, res, next) => {
+  res.sendFile(__dirname + '/client/login.html');
+});
+
+app.get('/dashboard', middleware.isLoggedIn, (req, res, next) => {
+  res.sendFile(__dirname + '/client/dashboard.html');
+});
+
 app.get('/', (req, res, next) => {
   res.sendFile(__dirname + '/client/index.html');
-});
-
-app.get('/register',(req, res, next) =>{
-  res.sendFile(__dirname + "/client/register.html")
-});
-
-app.get('/login',(req, res, next) =>{
-  res.sendFile(__dirname + "/client/login.html")
-});
-
-app.get('/register',(req, res, next) =>{
-  res.sendFile(__dirname + "/client/register.html")
-});
-
-app.get('/login',(req, res, next) =>{
-  res.sendFile(__dirname + "/client/login.html")
-});
-
-app.get("/dashboard",middleware.isLoggedIn,(req,res,next)=>{
-  res.sendFile(__dirname + "/client/dashboard.html");
 });
 
 app.listen(PORT, () => {
